+++ src/lang/interpreter.cpp
@@ -92,6 +92,7 @@ Number Interpreter::calc(const Operation::Type type, const Number& target,
     case Operation::Type::LPB:
     case Operation::Type::LPE:
     case Operation::Type::CLR:
+    case Operation::Type::SOR:
     case Operation::Type::SEQ:
       Log::get().error(
           "non-arithmetic operation: " + Operation::Metadata::get(type).name,
@@ -216,6 +217,12 @@ size_t Interpreter::run(const Program& p, Memory& mem) {
         length = get(op.source, mem).asInt();
         start = get(op.target, mem, true).asInt();
         mem.clear(start, length);
+        break;
+      }
+      case Operation::Type::SOR: {
+        length = get(op.source, mem).asInt();
+        start = get(op.target, mem, true).asInt();
+        mem.sort(start, length);
         break;
       }
       case Operation::Type::DBG: {