+++ src/lang/optimizer.cpp
@@ -216,7 +216,7 @@ bool Optimizer::mergeOps(Program &p) const {
           o2.source.type == Operand::Type::CONSTANT) {
         // first operation writing target?
         if (Operation::Metadata::get(o1.type).is_writing_target &&
-            o1.type != Operation::Type::CLR) {
+            !ProgramUtil::isWritingRange(o1.type)) {
           // second mov overwrites first operation
           o1 = o2;
           do_merge = true;
@@ -334,6 +334,7 @@ bool Optimizer::simplifyOperations(Program &p) const {
       case Operation::Type::LPB:
       case Operation::Type::LPE:
       case Operation::Type::CLR:
+      case Operation::Type::SOR:
       case Operation::Type::SEQ:
         can_simplify = false;
         break;
@@ -606,7 +607,7 @@ bool Optimizer::doPartialEval(Program &p, size_t op_index,
       // remove values from cells that are modified in the loop
       auto loop = ProgramUtil::getEnclosingLoop(p, op_index);
       for (int64_t i = loop.first + 1; i < loop.second; i++) {
-        if (p.ops[i].type == Operation::Type::CLR ||
+        if (ProgramUtil::isWritingRange(p.ops[i].type) ||
             ProgramUtil::hasIndirectOperand(p.ops[i])) {
           values.clear();
           break;
@@ -618,7 +619,8 @@ bool Optimizer::doPartialEval(Program &p, size_t op_index,
       }
       return false;
     }
-    case Operation::Type::CLR: {
+    case Operation::Type::CLR:
+    case Operation::Type::SOR: {
       values.clear();
       return false;
     }