+++ src/lang/program.cpp
@@ -4,7 +4,7 @@
 
 #include "lang/number.hpp"
 
-const std::array<Operation::Type, 23> Operation::Types = {
+const std::array<Operation::Type, 24> Operation::Types = {
     Operation::Type::NOP, Operation::Type::MOV, Operation::Type::ADD,
     Operation::Type::SUB, Operation::Type::TRN, Operation::Type::MUL,
     Operation::Type::DIV, Operation::Type::DIF, Operation::Type::MOD,
@@ -12,7 +12,7 @@ const std::array<Operation::Type, 23> Operation::Types = {
     Operation::Type::EQU, Operation::Type::NEQ, Operation::Type::LEQ,
     Operation::Type::GEQ, Operation::Type::MIN, Operation::Type::MAX,
     Operation::Type::LPB, Operation::Type::LPE, Operation::Type::CLR,
-    Operation::Type::SEQ, Operation::Type::DBG,
+    Operation::Type::SOR, Operation::Type::SEQ, Operation::Type::DBG,
 };
 
 const Operation::Metadata& Operation::Metadata::get(Type t) {
@@ -58,6 +58,8 @@ const Operation::Metadata& Operation::Metadata::get(Type t) {
       Operation::Type::LPE, "lpe", 'e', 0, true, false, false};
   static Operation::Metadata clr{
       Operation::Type::CLR, "clr", 'r', 2, true, false, true};
+  static Operation::Metadata sor{
+      Operation::Type::SOR, "sor", 'y', 2, true, true, true};
   static Operation::Metadata seq{
       Operation::Type::SEQ, "seq", 'q', 2, true, true, true};
   static Operation::Metadata dbg{
@@ -105,6 +107,8 @@ const Operation::Metadata& Operation::Metadata::get(Type t) {
       return lpe;
     case Operation::Type::CLR:
       return clr;
+    case Operation::Type::SOR:
+      return sor;
     case Operation::Type::SEQ:
       return seq;
     case Operation::Type::DBG: