+++ CHANGELOG.md
@@ -4,7 +4,8 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ### Features
 
-* Add operations: `clr`,`neq`,`leq`,`geq`
+* Add comparison operations: `neq`,`leq`,`geq`,`sor`
+* Add region operations: `clr` (clear), `sor` (sort)
 
 ### Enhancements